Function pdqsort::recurse

fn recurse<'a, T, F>(v: &'a mut [T],
                     is_less: &mut F,
                     pred: Option<&'a T>,
                     limit: usize) where F: FnMut(&T, &T) -> bool

Sorts v recursively.

If the slice had a predecessor in the original array, it is specified as pred.

limit is the number of allowed imbalanced partitions before switching to heapsort. If zero, this function will immediately switch to heapsort.